chore: Renamed EmbeddingFunctions.py to embedding_functions.py to follow snakecase convention
